Adding another option to deploy the code through script alongside Jupyter Notebook. This would enable the users to take this sample and extend for their need seamlessly and help automation.
